Revert "Unfreezing eraser rotation"
This reverts commit 29d8f5555526a17752041a846d13f203a2c54832.
