+++ Assets/Canvas/Canvas.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Canvas : MonoBehaviour
+{
+    public Texture2D texture;
+    public Vector2 textureSize = new Vector2(2048, 2048);
+
+    void Start()
+    {
+        var r = GetComponent<Renderer>();
+        texture = new Texture2D((int)textureSize.x, (int)textureSize.y);
+        r.material.mainTexture = texture;
+    }
+}
