+++ Assets/Scenes/BasicScene.unity
@@ -1989,7 +1989,7 @@ Rigidbody:
   m_UseGravity: 1
   m_IsKinematic: 0
   m_Interpolate: 0
-  m_Constraints: 0
+  m_Constraints: 112
   m_CollisionDetection: 2
 --- !u!114 &1095944871
 MonoBehaviour:
