+++ Assets/Scripts/CanvasEraser.cs
@@ -1,5 +1,4 @@
 using System.Linq;
-using Canvas;
 using UnityEngine;
 
 public class CanvasEraser : MonoBehaviour
@@ -13,7 +12,7 @@ public class CanvasEraser : MonoBehaviour
     private float _tipHeight;
 
     private RaycastHit _touch;
-    private CanvasScript canvasScript;
+    private Canvas _canvas;
     private Vector2 _touchPos, _lastTouchPos;
     private bool _touchedLastFrame;
     private Quaternion _lastTouchRot;
@@ -34,32 +33,32 @@ public class CanvasEraser : MonoBehaviour
     {
         if (Physics.Raycast(_tip.position, transform.up, out _touch, _tipHeight))
         {
-            if (_touch.transform.CompareTag("CanvasScript"))
+            if (_touch.transform.CompareTag("Canvas"))
             {
-                if (canvasScript == null)
+                if (_canvas == null)
                 {
-                    canvasScript = _touch.transform.GetComponent<CanvasScript>();
+                    _canvas = _touch.transform.GetComponent<Canvas>();
                 }
 
                 _touchPos = new Vector2(_touch.textureCoord.x, _touch.textureCoord.y);
 
-                var x = (int)(_touchPos.x * canvasScript.textureSize.x - (_eraserSize / 2));
-                var y = (int)(_touchPos.y * canvasScript.textureSize.y - (_eraserSize / 2));
+                var x = (int)(_touchPos.x * _canvas.textureSize.x - (_eraserSize / 2));
+                var y = (int)(_touchPos.y * _canvas.textureSize.y - (_eraserSize / 2));
 
-                if (y < 0 || y > canvasScript.textureSize.y || x < 0 || x > canvasScript.textureSize.x) return;
+                if (y < 0 || y > _canvas.textureSize.y || x < 0 || x > _canvas.textureSize.x) return;
 
                 if (_touchedLastFrame)
                 {
-                    canvasScript.texture.SetPixels(x, y, _eraserSize, _eraserSize, _clearColors);
+                    _canvas.texture.SetPixels(x, y, _eraserSize, _eraserSize, _clearColors);
 
                     for (float f = 0.01f; f < 1.00f; f += 0.01f)
                     {
                         var lerpX = (int)Mathf.Lerp(_lastTouchPos.x, x, f);
                         var lerpY = (int)Mathf.Lerp(_lastTouchPos.y, y, f);
-                        canvasScript.texture.SetPixels(lerpX, lerpY, _eraserSize, _eraserSize, _clearColors);
+                        _canvas.texture.SetPixels(lerpX, lerpY, _eraserSize, _eraserSize, _clearColors);
                     }
 
-                    canvasScript.texture.Apply();
+                    _canvas.texture.Apply();
                 }
 
                 _lastTouchPos = new Vector2(x, y);
@@ -69,7 +68,7 @@ public class CanvasEraser : MonoBehaviour
             }
         }
 
-        canvasScript = null;
+        _canvas = null;
         _touchedLastFrame = false;
     }
 }
